Issue #SB-19007 feat: Player orientation change to landscape for all non-DIAL pages on mobile web
