+++ src/app/client/src/app/modules/learn/components/course-consumption/course-player/course-player.component.html
@@ -112,7 +112,7 @@
                                 </div>
                                 <div class="ui grid">
                                     <div class="twelve wide column" [appTelemetryImpression]="telemetryContentImpression">
-                                        <app-player [contentProgressEvents$]="contentProgressEvents$" (assessmentEvents)="assessmentEvents($event)" (questionScoreSubmitEvents)="questionScoreSubmitEvents($event)" class="content-player" [playerConfig] ="playerConfig" [playerOption] = "playerOption"></app-player>
+                                        <app-player [telemetryObject]="contentInteractObject" [pageId]="pageId" [overlayImagePath]="overlayImagePath" [contentProgressEvents$]="contentProgressEvents$" (assessmentEvents)="assessmentEvents($event)" (questionScoreSubmitEvents)="questionScoreSubmitEvents($event)" class="content-player" [playerConfig] ="playerConfig" [playerOption] = "playerOption"></app-player>
                                     </div>
                                 </div>
                             </div>
