+++ src/app/client/src/app/modules/learn/components/course-consumption/course-player/course-player.component.ts
@@ -113,6 +113,8 @@ export class CoursePlayerComponent implements OnInit, OnDestroy {
   public unsubscribe = new Subject<void>();
   public contentProgressEvents$ = new Subject();
   playerOption: any;
+  overlayImagePath: string;
+  pageId: string;
   constructor(public activatedRoute: ActivatedRoute, private configService: ConfigService,
     private courseConsumptionService: CourseConsumptionService, public windowScrollService: WindowScrollService,
     public router: Router, public navigationHelperService: NavigationHelperService, private userService: UserService,
@@ -128,6 +130,7 @@ export class CoursePlayerComponent implements OnInit, OnDestroy {
     };
   }
   ngOnInit() {
+    this.pageId = this.activatedRoute.snapshot.data.telemetry.pageid;
     merge(this.activatedRoute.params.pipe(first(),
     mergeMap(({ courseId, batchId, courseStatus }) => {
       this.courseId = courseId;
@@ -151,6 +154,7 @@ export class CoursePlayerComponent implements OnInit, OnDestroy {
     .subscribe(({ courseHierarchy, enrolledBatchDetails, contentProgressEvent }: any) => {
        if (!contentProgressEvent) {
         this.courseHierarchy = courseHierarchy;
+        this.overlayImagePath = _.get(courseHierarchy, 'appIcon');
         this.contributions = _.join(_.map(this.courseHierarchy.contentCredits, 'name'));
         this.courseInteractObject = {
           id: this.courseHierarchy.identifier,
