+++ src/app/client/src/app/modules/player-helper/components/player/player.component.html
@@ -1,4 +1,20 @@
-<iframe #contentIframe id="contentPlayer" class="contentViewerIframeShadow" name="contentPlayer"></iframe>
+<!-- Mobile browser player-->
+<div *ngIf="isMobileOrTab" >
+    <div data-ratio="16:9" class="player-thumbnail" appTelemetryInteract [telemetryInteractObject]="telemetryObject" [telemetryInteractEdata]="loadPlayerInteractEdata"
+    [ngStyle]="{'background': 'url(' + overlayImagePath + ') center/cover no-repeat '}" *ngIf="playerThumbnail" id="contentPlayer" (click)="enablePlayer(false)">
+    <img class="play-icon" src="assets/images/play_circle.svg">
+  </div>
+  <div id="playerFullscreen" *ngIf="!playerThumbnail" > 
+    <img appTelemetryInteract [telemetryInteractObject]="telemetryInteractObject" [telemetryInteractEdata]="closeButtonInteractEdata" *ngIf="showCloseButton" (click)="closeFullscreen()" class="close-button" src="assets/images/close-Black.svg" alt="">
+    <iframe #contentIframe id="contentPlayer" class="contentViewerIframeShadow" name="contentPlayer"></iframe>
+  </div>
+</div>
+<!-- Portal player-->
+<div *ngIf="!isMobileOrTab">
+    <iframe #contentIframe id="contentPlayer" class="contentViewerIframeShadow" name="contentPlayer"></iframe>
+</div>
+
+    <!--Rating popup -->
 <app-content-rating #modal *ngIf="contentRatingModal &&  playerOption?.showContentRating && playerConfig"
     [contentData]="playerConfig.metadata"></app-content-rating>
 <!-- <button class="sb-btn sb-btn-primary sb-btn-normal pull-right view-fullscreen"[ngClass]="{'d-none': !viewFullscreenBtn}" (click)="viewInFullscreen()" appTelemetryInteract [telemetryInteractEdata]="viewFullScreenIntractEdata" [telemetryInteractObject]="viewFullScreenIntractObject"> <i class="expand icon"></i> View Full Screen</button> -->
