+++ src/app/client/src/app/modules/player-helper/components/player/player.component.spec.ts
@@ -1,5 +1,5 @@
 import { SharedModule } from '@sunbird/shared';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { async, ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
 import { PlayerComponent } from './player.component';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { RouterTestingModule } from '@angular/router/testing';
@@ -52,21 +52,19 @@ describe('PlayerComponent', () => {
     fixture = TestBed.createComponent(PlayerComponent);
     component = fixture.componentInstance;
     component.contentProgressEvents$ = new Subject();
+    component.contentIframe = { nativeElement: { contentWindow: { EkstepRendererAPI: { getCurrentStageId: () => 'stageId' } } } };
   });
 
-  it('should emit "START"', () => {
+  it('should emit "START"', fakeAsync(() => {
     let contentProgressEvent;
     component.contentProgressEvents$.subscribe((data) => {
       contentProgressEvent = data;
     });
-    component.contentIframe.nativeElement.contentWindow.EkstepRendererAPI = {
-      getCurrentStageId: () => 'stageId'
-    };
     spyOn(component, 'emitSceneChangeEvent').and.callFake(() => 'called');
     component.generateContentReadEvent(startEvent);
     component.playerConfig = playerConfig;
     expect(contentProgressEvent).toBeDefined();
-  });
+  }));
 
   it('should emit "END" event', () => {
     let contentProgressEvent;
@@ -93,5 +91,118 @@ describe('PlayerComponent', () => {
     component.ngOnChanges();
     expect(component.contentRatingModal).toBeFalsy();
    });
+
+   describe('should rotate player', () => {
+    let mockDomElement;
+    beforeEach(() => {
+        mockDomElement = document.createElement('div');
+        mockDomElement.setAttribute('id', 'playerFullscreen');
+    });
+
+    it('should rotate player for a default chrome browser', fakeAsync(() => {
+      spyOn(document, 'querySelector').and.returnValue(mockDomElement);
+      spyOn(screen.orientation, 'lock');
+      component.rotatePlayer();
+      tick(100);
+      expect(component.showCloseButton).toBe(true);
+      expect(screen.orientation.lock).toHaveBeenCalledWith('landscape');
+    }));
+
+    it('should rotate player for mozilla browser', fakeAsync(() => {
+      mockDomElement.requestFullscreen = undefined;
+      mockDomElement.mozRequestFullScreen = () => {};
+      spyOn(document, 'querySelector').and.returnValue(mockDomElement);
+      spyOn(screen.orientation, 'lock');
+      component.rotatePlayer();
+      tick(100);
+      expect(component.showCloseButton).toBe(true);
+      expect(screen.orientation.lock).toHaveBeenCalledWith('landscape');
+    }));
+
+    it('should rotate player for webkit browser', fakeAsync(() => {
+      mockDomElement.requestFullscreen = undefined;
+      mockDomElement.mozRequestFullScreen = undefined;
+      mockDomElement.webkitRequestFullscreen = () => {};
+      spyOn(document, 'querySelector').and.returnValue(mockDomElement);
+      spyOn(screen.orientation, 'lock');
+      component.rotatePlayer();
+      tick(100);
+      expect(component.showCloseButton).toBe(true);
+      expect(screen.orientation.lock).toHaveBeenCalledWith('landscape');
+    }));
+
+    it('should rotate player ms browser', fakeAsync(() => {
+      mockDomElement.requestFullscreen = undefined;
+      mockDomElement.mozRequestFullScreen = undefined;
+      mockDomElement.webkitRequestFullscreen = undefined;
+      mockDomElement.msRequestFullscreen = () => {};
+      spyOn(document, 'querySelector').and.returnValue(mockDomElement);
+      spyOn(screen.orientation, 'lock');
+      component.rotatePlayer();
+      tick(100);
+      expect(component.showCloseButton).toBe(true);
+      expect(screen.orientation.lock).toHaveBeenCalledWith('landscape');
+    }));
+  });
+
+  describe('should close the browser fullscreen mode', () => {
+    it('should close player fullscreen for default chrome browser', () => {
+      component.isSingleContent = true;
+      component.closeFullscreen();
+      expect(component.showCloseButton).toBe(false);
+      expect(component.playerThumbnail).toBe(true);
+    });
+
+    it('should close player fullscreen for mozilla browser', () => {
+      document['exitFullscreen'] = undefined;
+      document['mozCancelFullScreen'] = () => {};
+      component.isSingleContent = true;
+      component.closeFullscreen();
+      expect(component.showCloseButton).toBe(false);
+      expect(component.playerThumbnail).toBe(true);
+    });
+
+    it('should close player fullscreen for webkit browser ', () => {
+      document['exitFullscreen'] = undefined;
+      document['mozCancelFullScreen'] = undefined;
+      document['webkitExitFullscreen'] = () => {};
+      component.isSingleContent = true;
+      component.closeFullscreen();
+      expect(component.showCloseButton).toBe(false);
+      expect(component.playerThumbnail).toBe(true);
+    });
+
+    it('should close player fullscreen for ms browser ', () => {
+      document['exitFullscreen'] = undefined;
+      document['mozCancelFullScreen'] = undefined;
+      document['webkitExitFullscreen'] = undefined;
+      document['msExitFullscreen'] = () => {};
+      component.isSingleContent = true;
+      component.closeFullscreen();
+      expect(component.showCloseButton).toBe(false);
+      expect(component.playerThumbnail).toBe(true);
+    });
+  });
+
+  it('should load player on tap of play icon', () => {
+    spyOn(component, 'loadPlayer');
+    spyOn(component, 'rotatePlayer').and.stub();
+    component.enablePlayer(true);
+    expect(component.playerThumbnail).toBe(true);
+    expect(component.loadPlayer).toHaveBeenCalled();
+  });
+
+  it('should close player fullscreen ', () => {
+    component.isSingleContent = true;
+    component.closeFullscreen();
+    expect(component.showCloseButton).toBe(false);
+    expect(component.playerThumbnail).toBe(true);
+  });
+
+  it('should close player fullscreen and enable top div', () => {
+    component.isSingleContent = false;
+    component.closeFullscreen();
+    expect(component.showCloseButton).toBe(false);
+  });
 });
