+++ src/app/client/src/app/modules/player-helper/components/player/player.component.ts
@@ -1,5 +1,6 @@
 import { ConfigService, NavigationHelperService } from '@sunbird/shared';
-import { Component, AfterViewInit, ViewChild, ElementRef, Input, Output, EventEmitter, OnChanges } from '@angular/core';
+import { Component, AfterViewInit, ViewChild, ElementRef, Input, Output, EventEmitter,
+OnChanges, HostListener, OnInit } from '@angular/core';
 import * as _ from 'lodash-es';
 import { PlayerConfig } from '@sunbird/shared';
 import { environment } from '@sunbird/environment';
@@ -7,11 +8,14 @@ import { Router } from '@angular/router';
 import { ToasterService, ResourceService } from '@sunbird/shared';
 const OFFLINE_ARTIFACT_MIME_TYPES = ['application/epub', 'video/webm', 'video/mp4', 'application/pdf'];
 import { Subject } from 'rxjs';
+import { DeviceDetectorService } from 'ngx-device-detector';
+import { IInteractEventEdata } from '@sunbird/telemetry';
 @Component({
   selector: 'app-player',
-  templateUrl: './player.component.html'
+  templateUrl: './player.component.html',
+  styleUrls: ['./player.component.scss']
 })
-export class PlayerComponent implements AfterViewInit, OnChanges {
+export class PlayerComponent implements OnInit, AfterViewInit, OnChanges {
   @Input() playerConfig: PlayerConfig;
   @Output() assessmentEvents = new EventEmitter<any>();
   @Output() questionScoreSubmitEvents = new EventEmitter<any>();
@@ -28,18 +32,46 @@ export class PlayerComponent implements AfterViewInit, OnChanges {
   CONSTANT = {
     ACCESSEVENT: 'renderer:question:submitscore'
   };
+
+  @Input() overlayImagePath: string;
+  @Input() isSingleContent: boolean;
+  @Input() telemetryObject: {};
+  @Input() pageId: string;
+  @Output() closePlayerEvent = new EventEmitter<any>();
+  isMobileOrTab: boolean;
+  playerThumbnail = true;
+  showCloseButton = false;
+  closeButtonInteractEdata: IInteractEventEdata;
+  loadPlayerInteractEdata: IInteractEventEdata;
   /**
  * Dom element reference of contentRatingModal
  */
   @ViewChild('modal') modal;
   constructor(public configService: ConfigService, public router: Router, private toasterService: ToasterService,
-    public resourceService: ResourceService, public navigationHelperService: NavigationHelperService) {
+    public resourceService: ResourceService, public navigationHelperService: NavigationHelperService,
+    private deviceDetectorService: DeviceDetectorService) {
     this.buildNumber = (<HTMLInputElement>document.getElementById('buildNumber'))
       ? (<HTMLInputElement>document.getElementById('buildNumber')).value : '1.0';
     this.previewCdnUrl = (<HTMLInputElement>document.getElementById('previewCdnUrl'))
       ? (<HTMLInputElement>document.getElementById('previewCdnUrl')).value : undefined;
     this.isCdnWorking = (<HTMLInputElement>document.getElementById('cdnWorking'))
       ? (<HTMLInputElement>document.getElementById('cdnWorking')).value : 'no';
+  }
+
+  @HostListener('window:orientationchange', ['$event'])
+  public handleOrientationChange() {
+    const screenType = _.get(screen, 'orientation.type');
+      if ( screenType === 'portrait-primary' || screenType === 'portrait-secondary' ) {
+        this.closeFullscreen();
+      }
+  }
+
+  ngOnInit() {
+    this.isMobileOrTab = this.deviceDetectorService.isMobile() || this.deviceDetectorService.isTablet();
+    if (this.isSingleContent === false) {
+      this.playerThumbnail = false;
+    }
+    this.setTelemetryData();
   }
   /**
    * loadPlayer method will be called
@@ -118,6 +150,9 @@ export class PlayerComponent implements AfterViewInit, OnChanges {
       return;
     }
     this.loadDefaultPlayer();
+    if (this.isMobileOrTab) {
+      this.rotatePlayer();
+    }
   }
   /**
    * Adjust player height after load
@@ -168,5 +203,72 @@ export class PlayerComponent implements AfterViewInit, OnChanges {
         this.modal.showContentRatingModal = true;
       }
     }
+  }
+
+  /**
+   * this method will handle play button click and turn the player into landscape
+   */
+  enablePlayer(mode: boolean) {
+    this.playerThumbnail = mode;
+    this.loadPlayer();
+  }
+
+  /** this method checks for the browser capability to be fullscreen via if-else ladder
+   * if match found, it will turn the player along will be close button into fullscreen and then
+   * rotate it to landscape mode
+   */
+  rotatePlayer() {
+    setTimeout(() => {
+      const playVideo: any = document.querySelector('#playerFullscreen');
+      try {
+        if (playVideo.requestFullscreen) {
+          playVideo.requestFullscreen();
+        } else if (playVideo.mozRequestFullScreen) { /* Firefox */
+          playVideo.mozRequestFullScreen();
+        } else if (playVideo.webkitRequestFullscreen) { /* Chrome, Safari and Opera */
+          playVideo.webkitRequestFullscreen();
+        } else if (playVideo.msRequestFullscreen) { /* IE/Edge */
+          playVideo.msRequestFullscreen();
+        }
+        screen.orientation.lock('landscape');
+        this.showCloseButton = true;
+      } catch (error) {}
+    });
+  }
+
+  /** when user clicks on close button
+   * this method will let the player to exit from fullscreen mode and
+   * 1. video thumbnail will be shown for single content
+   * 2. content-details page will be shown ( for multi-result dial-code search flow)
+   */
+  closeFullscreen() {
+    this.showCloseButton = false;
+    /** to exit the fullscreen mode */
+    if (document['exitFullscreen']) {
+      document['exitFullscreen']();
+    } else if (document['mozCancelFullScreen']) { /* Firefox */
+      document['mozCancelFullScreen']();
+    } else if (document['webkitExitFullscreen']) { /* Chrome, Safari and Opera */
+      document['webkitExitFullscreen']();
+    } else if (document['msExitFullscreen']) { /* IE/Edge */
+      document['msExitFullscreen']();
+    }
+     /** to change the view of the content-details page */
+    this.playerThumbnail = true;
+    this.closePlayerEvent.emit();
+  }
+
+  setTelemetryData() {
+    this.closeButtonInteractEdata = {
+      id: 'player-close-button',
+      type: 'click',
+      pageid: this.pageId
+    };
+
+    this.loadPlayerInteractEdata = {
+      id: 'play-button',
+      type: 'click',
+      pageid: this.pageId
+    };
   }
 }
