+++ src/app/client/src/app/modules/public/module/player/components/public-collection-player/public-collection-player.component.html
@@ -27,7 +27,7 @@
       <!--/Content Title-->
 
       <!--Content Player-->
-      <app-player class="content-player" [playerOption] ="playerOption" [playerConfig]="playerConfig | async"></app-player>
+      <app-player class="content-player" [telemetryObject]="playerTelemetryInteractObject" [pageId]="pageId" [overlayImagePath]="overlayImagePath" [playerOption] ="playerOption" [playerConfig]="playerConfig | async"></app-player>
       <!--/Content Player-->
 
       <!--Content Navigation-->
