+++ src/app/client/src/app/modules/public/module/player/components/public-content-player/public-content-player.component.html
@@ -51,15 +51,8 @@
         <!--/Warning message for external content-->
 
         <!--Player-->
-        <div data-ratio="16:9" class="player-thumbnail" appTelemetryInteract [telemetryInteractObject]="telemetryInteractObject" [telemetryInteractEdata]="loadPlayerInteractEdata"
-        [ngStyle]="{'background': 'url(' + overlayImagePath + ') center/cover no-repeat '}" *ngIf="playerThumbnail" id="contentPlayer" (click)="enablePlayer(false)">
-        <img class="play-icon" src="assets/images/play_circle.svg">
-      </div>
-      <div id="playerFullscreen" *ngIf="!playerThumbnail" > 
-          <img appTelemetryInteract [telemetryInteractObject]="telemetryInteractObject" [telemetryInteractEdata]="closeButtonInteractEdata" *ngIf="showCloseButton" (click)="closeFullscreen()" class="close-button" src="assets/images/close-Black.svg" alt="">
-          <app-player class="content-player" [playerOption]="playerOption"  [playerConfig]="playerConfig"></app-player>
-      </div>
-      <!-- / Player-->
+        <app-player class="content-player" [telemetryObject]="telemetryInteractObject" [pageId]="'public'" [overlayImagePath]="overlayImagePath" [playerOption]="playerOption"  [playerConfig]="playerConfig"></app-player>
+        <!-- / Player-->
 
         <!--Meta Data of content-->
         <div class="ui stackable grid">
@@ -81,8 +74,7 @@
     <div class="twelve wide column" >
         <div *ngIf="(popupControlService.checkPopupStatus | async) else loading">
             <div class="player-fullscreen" id="playerFullscreen"> 
-              <img appTelemetryInteract [telemetryInteractObject]="telemetryInteractObject" [telemetryInteractEdata]="closeButtonInteractEdata" *ngIf="showCloseButton" (click)="closeFullscreen()" class="close-button" src="assets//images/close-Black.svg" alt="">
-              <app-player class="content-player content-player-landscape" [playerOption]="playerOption"  [playerConfig]="playerConfig"></app-player>
+              <app-player class="content-player content-player-landscape" (closePlayerEvent)="closePlayer()" [pageId]="'public'" [isSingleContent]="isSingleContent" [overlayImagePath]="overlayImagePath" [playerOption]="playerOption"  [playerConfig]="playerConfig"></app-player>
             </div>
             <ng-template #loading>
               <app-loader></app-loader>
