+++ src/app/client/src/app/modules/resource/modules/player/components/content-player/content-player.component.html
@@ -52,7 +52,7 @@
                             <p>{{resourceService.messages.imsg.m0036}}</p>
                         </div>
                     </div>
-                    <app-player   class="content-player" [playerOption] ="playerOption" [playerConfig]="playerConfig"></app-player>
+                    <app-player  [telemetryObject]="objectInteract" [pageId]="'content-player'"[overlayImagePath]="contentData.appIcon"  class="content-player" [playerOption] ="playerOption" [playerConfig]="playerConfig"></app-player>
                     <div class="twelve wide column contentMetaSection">
                         <div class="ui stackable grid mt-30">
                             <div class="row">
